requirements.txt
torch >= 2.2.0             # all
jupyterlab >= 4.0          # all
tiktoken >= 0.5.1          # ch02; ch04; ch05
matplotlib >= 3.7.1        # ch04; ch06; ch07
tensorflow >= 2.16.0       # ch05; ch06; ch07
tqdm >= 4.66.1             # ch05; ch07
numpy >= 1.26, < 2.1       # dependency of several other libraries like torch and pandas
pandas >= 2.2.1            # ch06
psutil >= 5.9.5            # ch07; already installed automatically as dependency of torch
pytorch-lightning>=2.5.0